refactor(e2e): fold deploy + nested contract tests into automine
Fold DeployTest and NestedContractTest into AutomineTestContext: registerContract
and the StatefulContractCtorArgs/ContractArtifactClass aliases move onto the base,
and applyManual becomes applyManualParentChild. Move the deploy tests under
automine/contracts/deploy/, the nested tests under automine/contracts/nested/, and
e2e_storage_proof under automine/contracts/ (with its fixtures). Files with more
than one top-level it gain the .parallel suffix.

Adds the automine/contracts globs to both bootstrap.sh test-list arrays, extends
the compat flatten branch to cover src/automine/*, repoints the storage_proof_fetcher
eslint override, the snapshot-regen script, and the protocol-contracts fixture
provenance comment.
