+++ yarn-project/end-to-end/src/automine/contracts/fixtures/storage_proof_fixture.ts
