+++ yarn-project/end-to-end/src/automine/contracts/nested/manual_private_call.test.ts
@@ -1,16 +1,14 @@
-import { AUTOMINE_E2E_OPTS } from '../fixtures/fixtures.js';
-import { NestedContractTest } from './nested_contract_test.js';
+import { AutomineTestContext } from '../../automine_test_context.js';
 
 // Tests a nested private call from ParentContract into ChildContract's value() function.
-// NestedContractTest wraps setup(0, { ...AUTOMINE_E2E_OPTS, fundSponsoredFPC, skipAccountDeployment })
-// with 1 public-deployed account. applyManual() deploys Parent and Child contracts in beforeAll.
-describe('e2e_nested_contract manual', () => {
-  const t = new NestedContractTest('manual');
+// Runs on a single account. applyManualParentChild() deploys Parent and Child contracts in beforeAll.
+describe('automine/contracts/nested/manual_private_call', () => {
+  const t = new AutomineTestContext();
   let { parentContract, childContract, defaultAccountAddress } = t;
 
   beforeAll(async () => {
-    await t.setup({ ...AUTOMINE_E2E_OPTS });
-    await t.applyManual();
+    await t.setup();
+    await t.applyManualParentChild();
     ({ parentContract, childContract, defaultAccountAddress } = t);
   });
