+++ yarn-project/end-to-end/src/automine/contracts/storage_proof.test.ts
@@ -2,26 +2,25 @@ import { StorageProofTestContract } from '@aztec/noir-test-contracts.js/StorageP
 
 import { jest } from '@jest/globals';
 
-import { AUTOMINE_E2E_OPTS } from '../fixtures/fixtures.js';
-import { type EndToEndContext, setup, teardown } from '../fixtures/setup.js';
+import { AutomineTestContext } from '../automine_test_context.js';
 import { buildStorageProofCapsules, loadStorageProofArgs } from './fixtures/storage_proof_fixture.js';
 
 jest.setTimeout(300_000);
 
 // Tests that a Noir contract can verify an Ethereum storage proof (MPT proof) via oracle capsules.
-// Plain setup(1, { ...AUTOMINE_E2E_OPTS }) with 1 account. Deploys StorageProofTestContract, then
-// loads pre-computed proof args from fixtures/storage_proof.json and verifies on-chain.
-describe('Storage proof', () => {
-  let ctx: EndToEndContext;
+// Deploys StorageProofTestContract, then loads pre-computed proof args from fixtures/storage_proof.json
+// and verifies on-chain.
+describe('automine/contracts/storage_proof', () => {
+  const t = new AutomineTestContext();
   let contract: StorageProofTestContract;
 
   beforeAll(async () => {
-    ctx = await setup(1, { ...AUTOMINE_E2E_OPTS });
-    ({ contract } = await StorageProofTestContract.deploy(ctx.wallet).send({ from: ctx.accounts[0] }));
+    await t.setup({ numberOfAccounts: 1 });
+    ({ contract } = await StorageProofTestContract.deploy(t.wallet).send({ from: t.defaultAccountAddress }));
   });
 
   afterAll(async () => {
-    await teardown(ctx);
+    await t.teardown();
   });
 
   // Loads pre-computed ethAddress/slotKey/slotContents/root from storage_proof.json, builds oracle
@@ -31,12 +30,12 @@ describe('Storage proof', () => {
     const { ethAddress, slotKey, slotContents, root } = loadStorageProofArgs();
     const capsules = await buildStorageProofCapsules(contract.address);
 
-    ctx.logger.info('Sending storage proof TX...');
+    t.logger.info('Sending storage proof TX...');
 
     const { receipt } = await contract.methods
       .storage_proof(ethAddress, slotKey, slotContents, root)
       .with({ capsules })
-      .send({ from: ctx.accounts[0] });
+      .send({ from: t.defaultAccountAddress });
 
     expect(receipt.hasExecutionSucceeded()).toBe(true);
   });
