--- yarn-project/end-to-end/src/e2e_deploy_contract/deploy_test.ts
@@ -1,73 +1,0 @@
-import type { ContractArtifact } from '@aztec/aztec.js/abi';
-import type { AztecAddress } from '@aztec/aztec.js/addresses';
-import { type ContractBase, getContractInstanceFromInstantiationParams } from '@aztec/aztec.js/contracts';
-import { Fr } from '@aztec/aztec.js/fields';
-import type { PublicKeys } from '@aztec/aztec.js/keys';
-import { type Logger, createLogger } from '@aztec/aztec.js/log';
-import type { AztecNode } from '@aztec/aztec.js/node';
-import type { Wallet } from '@aztec/aztec.js/wallet';
-import type { StatefulTestContract } from '@aztec/noir-test-contracts.js/StatefulTest';
-import type { AztecNodeAdmin } from '@aztec/stdlib/interfaces/client';
-
-import { type EndToEndContext, type SetupOptions, setup, teardown } from '../fixtures/setup.js';
-import type { TestWallet } from '../test-wallet/test_wallet.js';
-
-export class DeployTest {
-  public context!: EndToEndContext;
-  public logger: Logger;
-  public wallet!: TestWallet;
-  public defaultAccountAddress!: AztecAddress;
-  public aztecNode!: AztecNode;
-  public aztecNodeAdmin!: AztecNodeAdmin;
-
-  constructor(testName: string) {
-    this.logger = createLogger(`e2e:e2e_deploy_contract:${testName}`);
-  }
-
-  async setup(opts: Partial<SetupOptions> = {}) {
-    this.logger.info('Setting up test environment');
-    this.context = await setup(1, {
-      ...opts,
-      fundSponsoredFPC: true,
-    });
-    this.aztecNode = this.context.aztecNodeService;
-    this.wallet = this.context.wallet;
-    this.aztecNodeAdmin = this.context.aztecNodeService;
-    this.defaultAccountAddress = this.context.accounts[0];
-    return this;
-  }
-
-  async teardown() {
-    await teardown(this.context);
-  }
-
-  async registerContract<T extends ContractBase>(
-    wallet: Wallet,
-    contractArtifact: ContractArtifactClass<T>,
-    opts: {
-      salt?: Fr;
-      publicKeys?: PublicKeys;
-      initArgs?: any[];
-      constructorName?: string;
-      deployer?: AztecAddress;
-    } = {},
-  ): Promise<T> {
-    const { salt, publicKeys, initArgs, constructorName, deployer } = opts;
-    const instance = await getContractInstanceFromInstantiationParams(contractArtifact.artifact, {
-      constructorArgs: initArgs ?? [],
-      constructorArtifact: constructorName,
-      salt: salt ?? Fr.random(),
-      publicKeys,
-      deployer,
-    });
-    await wallet.registerContract(instance, contractArtifact.artifact);
-    return contractArtifact.at(instance.address, wallet);
-  }
-}
-
-export type StatefulContractCtorArgs = Parameters<StatefulTestContract['methods']['constructor']>;
-
-export type ContractArtifactClass<T extends ContractBase> = {
-  at(address: AztecAddress, wallet: Wallet): T;
-  artifact: ContractArtifact;
-};
