+++ yarn-project/protocol-contracts/src/tests/fixtures.ts
@@ -2,13 +2,13 @@ import { readFileSync } from 'fs';
 import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
 
-// Generated from end-to-end/src/e2e_deploy_contract/contract_class_registration.test.ts with AZTEC_GENERATE_TEST_DATA=1
+// Generated from end-to-end/src/automine/contracts/deploy/contract_class_registration.test.ts with AZTEC_GENERATE_TEST_DATA=1
 export function getSampleContractClassPublishedEventPayload(): Buffer {
   const path = getPathToFixture('ContractClassPublishedEventData.hex');
   return Buffer.from(readFileSync(path).toString(), 'hex');
 }
 
-// Generated from end-to-end/src/e2e_deploy_contract/contract_class_registration.test.ts with AZTEC_GENERATE_TEST_DATA=1
+// Generated from end-to-end/src/automine/contracts/deploy/contract_class_registration.test.ts with AZTEC_GENERATE_TEST_DATA=1
 export function getSampleContractInstancePublishedEventPayload(): Buffer {
   const path = getPathToFixture('ContractInstancePublishedEventData.hex');
   return Buffer.from(readFileSync(path).toString(), 'hex');
