+++ yarn-project/update-snapshots.sh
@@ -7,7 +7,7 @@ yarn build:fast
 export AZTEC_GENERATE_TEST_DATA=1
 
 yarn workspace @aztec/end-to-end test integration_l1_publisher.test.ts
-yarn workspace @aztec/end-to-end test e2e_nested_contract -t 'performs nested calls'
+yarn workspace @aztec/end-to-end test automine/contracts/nested -t 'performs nested calls'
 
 # this test takes considerable resources to run since it fully proves blocks
 # only enable if needed
